+++ src/main/java/com/evacipated/cardcrawl/modthespire/lib/SpireField.java
@@ -18,6 +18,13 @@ public class SpireField<T>
         this.defaultValue = defaultValue;
     }
 
+    public SpireField(SpireField<T> originalSpireField)
+    {
+        if (originalSpireField != null) {
+            defaultValue = originalSpireField.defaultValue;
+        }
+    }
+
     public void initialize(Class clz, String fieldName) throws NoSuchFieldException
     {
         field = clz.getDeclaredField(fieldName);
@@ -31,6 +38,7 @@ public class SpireField<T>
 
     public T get(Object __instance)
     {
+        // This should never be called, but serves as a fallback
         try {
             return (T) field.get(__instance);
         } catch (IllegalAccessException e) {
@@ -41,6 +49,7 @@ public class SpireField<T>
 
     public void set(Object __instance, T value)
     {
+        // This should never be called, but serves as a fallback
         try {
             field.set(__instance, value);
         } catch (IllegalAccessException e) {